Now hfst-shuffle is disabled by default.
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk/hfst3@2302 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/shuffle-functionality.sh
@@ -1,4 +1,8 @@
 #!/bin/sh
+if ! [ -x ../../tools/src/hfst-shuffle ]; then
+    exit 0;
+fi
+
 for i in .sfst .ofst .foma; do
     if ! ../../tools/src/hfst-shuffle ab$i bc$i > TMP; then
 	rm TMP;
